# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "EventHubAPI | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/admin/events": {
      "get": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "Tags",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "Format",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EventStatus"
            }
          },
          {
            "name": "TagsConfirmed",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "AllDates",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEventResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminEvents"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/events/{eventId}": {
      "get": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/events/{eventId}/tags": {
      "put": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagIdsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TagIdsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TagIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/events/{eventId}/publish": {
      "post": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/events/{eventId}/unpublish": {
      "post": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/events/{eventId}": {
      "get": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventDto": {
        "required": [
          "title",
          "description",
          "eventDateTime",
          "location",
          "source",
          "deadline",
          "tagIds"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "eventDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "deadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "mainImg": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "EventResponse": {
        "required": [
          "id",
          "title",
          "description",
          "eventDateTime",
          "location",
          "source",
          "mainImg",
          "deadline",
          "eventStatus",
          "tagsConfirmed",
          "tagIds",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "eventDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "mainImg": {
            "type": [
              "null",
              "string"
            ]
          },
          "deadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "eventStatus": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "tagsConfirmed": {
            "type": "boolean"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EventStatus": {
        "type": "integer"
      },
...
      "PagedResultOfEventResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount",
          "hasNextPage"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "TagIdsRequest": {
        "required": [
          "tagIds"
        ],
        "type": "object",
        "properties": {
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}